Added Price Comparison
Added the code to compare pricings of each volume against each other and print a list of all the volumes available from each website that has the lowest price out of the websites. Also added the ability to input a series title and booktype through the cmd and some other minor bug fixes.